Move command to seperate tree
