+++ ydb/public/lib/ydb_cli/commands/ydb_dynamic_config.cpp
@@ -3,8 +3,6 @@
 #include <ydb-cpp-sdk/client/draft/ydb_dynamic_config.h>
 #include <ydb-cpp-sdk/client/config/config.h>
 #include <ydb/library/yaml_config/public/yaml_config.h>
-#include <util/system/fs.h>
-#include <util/folder/dirut.h>
 
 #include <openssl/sha.h>
 
@@ -14,10 +12,6 @@
 using namespace NKikimr;
 
 namespace NYdb::NConsoleClient::NDynamicConfig {
-
-constexpr const char* CONFIG_FILE_NAME = "config.yaml";
-constexpr const char* MAIN_CONFIG_FILE_NAME = "main.yaml";
-constexpr const char* STORAGE_CONFIG_FILE_NAME = "storage.yaml";
 
 TString WrapYaml(const TString& yaml) {
     auto doc = NFyaml::TDocument::Parse(yaml);
@@ -43,25 +37,6 @@ TString WrapStaticConfig(const TString& yaml) {
 
     return TString(newDoc.EmitToCharArray().get());
 }
-bool SaveConfig(const TString& config, const TString& configName, const TString& configDirPath) {
-    try {
-        TString tempPath = TStringBuilder() << configDirPath << "/temp_" << configName;
-        TString configPath = TStringBuilder() << configDirPath << "/" << configName;
-
-        {
-            TFileOutput tempFile(tempPath);
-            tempFile << config;
-            tempFile.Flush();
-        }
-
-        if (!NFs::Rename(tempPath, configPath)) {
-            return false;
-        }
-    } catch (const std::exception& e) {
-        return false;
-    }
-    return true;
-}
 
 TCommandConfig::TCommandConfig(
         TCommandFlagsOverrides commandFlagsOverrides,
@@ -73,7 +48,6 @@ TCommandConfig::TCommandConfig(
     AddCommand(std::make_unique<TCommandConfigReplace>(allowEmptyDatabase));
     AddCommand(std::make_unique<TCommandConfigResolve>());
     AddCommand(std::make_unique<TCommandGenerateDynamicConfig>(allowEmptyDatabase));
-    AddCommand(std::make_unique<TCommandConfigInit>());
 }
 
 TCommandConfig::TCommandConfig(bool allowEmptyDatabase)
@@ -784,79 +758,6 @@ int TCommandGenerateDynamicConfig::Run(TConfig& config) {
         Cout << "Startup config is already dynamic" << Endl;
     }
 
-    return EXIT_SUCCESS;
-}
-
-TCommandConfigInit::TCommandConfigInit()
-    : TYdbCommand("init", {}, "Initialize dynamic config")
-{
-}
-
-void TCommandConfigInit::Config(TConfig& config) {
-    TYdbCommand::Config(config);
-    config.Opts->AddLongOption('f', "from-config", "Path to the initial configuration file.")
-        .RequiredArgument("[config.yaml]").StoreResult(&ConfigYamlPath);
-    config.Opts->AddLongOption('d', "config-dir", "Path to the directory for storing configuration files.")
-        .Required().RequiredArgument("[directory]").StoreResult(&ConfigDirPath);
-    config.Opts->AddLongOption('s', "seed-node", "Endpoint of the seed node from which the configuration will be fetched.")
-        .RequiredArgument("[HOST:PORT]").StoreResult(&SeedNodeEndpoint);
-    config.Opts->MutuallyExclusive("from-config", "seed-node");
-    config.SetFreeArgsNum(0);
-    config.AllowEmptyDatabase = true;
-    config.AllowEmptyAddress = true;
-}
-
-int TCommandConfigInit::Run(TConfig& config) {
-
-    MakePathIfNotExist(ConfigDirPath.c_str());
-    TString configYaml;
-
-    if (SeedNodeEndpoint) {
-        config.Address = SeedNodeEndpoint;
-        auto driver = std::make_unique<NYdb::TDriver>(CreateDriver(config));
-        auto client = NYdb::NConfig::TConfigClient(*driver);
-
-        auto result = client.FetchAllConfigs().GetValueSync();
-        NStatusHelpers::ThrowOnErrorOrPrintIssues(result);
-        TString clusterConfig;
-        TString storageConfig;
-        for (const auto& entry : result.GetConfigs()) {
-            std::visit([&](auto&& arg) {
-                using T = std::decay_t<decltype(arg)>;
-                if constexpr (std::is_same_v<T, NYdb::NConfig::TMainConfigIdentity>) {
-                    clusterConfig = entry.Config;
-                } else if constexpr (std::is_same_v<T, NYdb::NConfig::TStorageConfigIdentity>) {
-                    storageConfig = entry.Config;
-                }
-            }, entry.Identity);
-        }
-        if (!clusterConfig.empty() && !storageConfig.empty()) {
-            if (SaveConfig(clusterConfig, MAIN_CONFIG_FILE_NAME, ConfigDirPath) &&
-                SaveConfig(storageConfig, STORAGE_CONFIG_FILE_NAME, ConfigDirPath)) {
-                Cout << "Initialized cluster and storage configs in " << ConfigDirPath << "/" << MAIN_CONFIG_FILE_NAME << " and " << STORAGE_CONFIG_FILE_NAME << Endl;
-                return EXIT_SUCCESS;
-            } else {
-                Cout << "Failed to initialize cluster config in " << ConfigDirPath << "/" << MAIN_CONFIG_FILE_NAME << Endl;
-                return EXIT_FAILURE;
-            }
-        }
-        else if (!clusterConfig.empty() && storageConfig.empty()) {
-            configYaml = clusterConfig;
-        }
-        else {
-            Cout << "No main config found on the seed node" << Endl;
-            return EXIT_FAILURE;
-        }
-    }
-    else {
-        configYaml = TFileInput(ConfigYamlPath).ReadAll();
-    }
-    if (SaveConfig(configYaml, CONFIG_FILE_NAME, ConfigDirPath)){
-        Cout << "Initialized cluster config in " << ConfigDirPath << "/" << CONFIG_FILE_NAME << Endl;
-    } else {
-        Cout << "Failed to initialize cluster config in " << ConfigDirPath << "/" << CONFIG_FILE_NAME << Endl;
-        return EXIT_FAILURE;
-    }
     return EXIT_SUCCESS;
 }
