+++ ydb/public/lib/ydb_cli/commands/ydb_dynamic_config.h
@@ -133,14 +133,4 @@ private:
     bool AllowEmptyDatabase = false;
 };
 
-class TCommandConfigInit : public TYdbCommand {
-public:
-    TCommandConfigInit();
-    void Config(TConfig& config) override;
-    int Run(TConfig& config) override;
-private:
-    TString ConfigYamlPath;
-    TString ConfigDirPath;
-    TString SeedNodeEndpoint;
-};
 } // namespace NYdb::NConsoleClient::NDynamicConfig
